# projects/integration-in-a-box/docs/api/harbourline.openapi.yaml
# Harbourline Inventory API, version 1.
# Written before any code, the way a vendor would hand it to a customer.
# Harbourline and Maple & Main are made-up companies. All data is synthetic.
openapi: 3.1.0
info:
  title: Harbourline Inventory API
  version: 1.0.0
  description: |
    Products, stock levels and orders for Harbourline customers.

    Milestone 1 has no authentication. OAuth 2.0 client credentials and
    signed webhooks are added in milestone 2.
servers:
  - url: http://localhost:4001
paths:
  /v1/health:
    get:
      summary: Is the API up?
      responses:
        "200": { description: OK }
  /v1/products:
    get:
      summary: List products, one page at a time
      description: |
        Results are ordered by SKU. Pass the `next_cursor` from one page as
        `cursor` to get the next page. `next_cursor` is null on the last page.
      parameters:
        - { name: limit, in: query, schema: { type: integer, minimum: 1, maximum: 100, default: 25 } }
        - { name: cursor, in: query, schema: { type: string } }
        - name: updated_since
          in: query
          description: Only products changed after this time (ISO 8601).
          schema: { type: string, format: date-time }
      responses:
        "200":
          description: A page of products
          content:
            application/json:
              schema:
                type: object
                required: [data, next_cursor]
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Product" } }
                  next_cursor: { type: [string, "null"] }
  /v1/products/{sku}:
    get:
      summary: One product
      parameters: [{ $ref: "#/components/parameters/Sku" }]
      responses:
        "200": { description: The product, content: { application/json: { schema: { $ref: "#/components/schemas/Product" } } } }
        "404": { $ref: "#/components/responses/Error" }
  /v1/products/{sku}/stock:
    post:
      summary: Adjust stock (a delivery arrives, or a count is corrected)
      parameters: [{ $ref: "#/components/parameters/Sku" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [change, reason]
              properties:
                change: { type: integer, description: "Units to add (positive) or remove (negative)." }
                reason: { type: string, enum: [delivery, count_correction, damage] }
      responses:
        "200": { description: Updated product, content: { application/json: { schema: { $ref: "#/components/schemas/Product" } } } }
        "404": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/orders:
    post:
      summary: Place an order, which takes stock
      description: |
        Send an `Idempotency-Key` header. Sending the same key again returns
        the original order instead of creating a second one, so a retry after
        a timeout can never double an order.
      parameters:
        - { name: Idempotency-Key, in: header, required: true, schema: { type: string, maxLength: 100 } }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [lines]
              properties:
                external_ref: { type: string, description: "The customer's own order id." }
                lines:
                  type: array
                  minItems: 1
                  items:
                    type: object
                    required: [sku, quantity]
                    properties:
                      sku: { type: string }
                      quantity: { type: integer, minimum: 1 }
      responses:
        "201": { description: Order created, content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } } }
        "200": { description: Same Idempotency-Key seen before; the original order, content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } } }
        "400": { $ref: "#/components/responses/Error" }
        "409": { $ref: "#/components/responses/Error" }
  /v1/orders/{id}:
    get:
      summary: One order
      parameters: [{ name: id, in: path, required: true, schema: { type: string } }]
      responses:
        "200": { description: The order, content: { application/json: { schema: { $ref: "#/components/schemas/Order" } } } }
        "404": { $ref: "#/components/responses/Error" }
components:
  parameters:
    Sku: { name: sku, in: path, required: true, schema: { type: string } }
  responses:
    Error:
      description: Something went wrong. `code` is stable; `message` is for people.
      content:
        application/json:
          schema:
            type: object
            required: [error]
            properties:
              error:
                type: object
                required: [code, message]
                properties:
                  code: { type: string, examples: [not_found, insufficient_stock, invalid_request] }
                  message: { type: string }
  schemas:
    Product:
      type: object
      required: [sku, name, unit_price_cents, stock_level, updated_at]
      properties:
        sku: { type: string, examples: [HB-1001] }
        name: { type: string }
        unit_price_cents: { type: integer }
        stock_level: { type: integer, minimum: 0 }
        updated_at: { type: string, format: date-time }
    Order:
      type: object
      required: [id, status, lines, created_at]
      properties:
        id: { type: string }
        external_ref: { type: [string, "null"] }
        status: { type: string, enum: [accepted] }
        lines:
          type: array
          items: { type: object, properties: { sku: { type: string }, quantity: { type: integer } } }
        created_at: { type: string, format: date-time }